✨ Feat: 검색 라우팅 처리 커스텀 훅으로 분리
